+++ src/test/java/ee/sk/test/smartid/integration/ReadmeTest.java
@@ -1,23 +1,6 @@
 package ee.sk.test.smartid.integration;
 
-import static java.util.Arrays.asList;
-
-import java.io.InputStream;
-import java.security.KeyStore;
-import java.security.cert.X509Certificate;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.concurrent.TimeUnit;
-
-import ee.sk.smartid.AuthenticationHash;
-import ee.sk.smartid.AuthenticationIdentity;
-import ee.sk.smartid.AuthenticationResponseValidator;
-import ee.sk.smartid.HashType;
-import ee.sk.smartid.SignableHash;
-import ee.sk.smartid.SmartIdAuthenticationResponse;
-import ee.sk.smartid.SmartIdCertificate;
-import ee.sk.smartid.SmartIdClient;
-import ee.sk.smartid.SmartIdSignature;
+import ee.sk.smartid.*;
 import ee.sk.smartid.exception.UnprocessableSmartIdResponseException;
 import ee.sk.smartid.exception.permanent.SmartIdClientException;
 import ee.sk.smartid.exception.useraccount.RequiredInteractionNotSupportedByAppException;
@@ -34,6 +17,17 @@ import org.junit.Before;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.io.InputStream;
+import java.security.KeyStore;
+import java.security.cert.X509Certificate;
+import java.time.LocalDate;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+
+import static java.util.Arrays.asList;
 
 /**
  * These tests contain snippets used in Readme.md
@@ -175,7 +169,7 @@ It is recommended to read trusted certificates from a file.
         SemanticsIdentifier semanticsIdentifier = new SemanticsIdentifier(
                 SemanticsIdentifier.IdentityType.PNO, // 3 character identity type (PAS-passport, IDC-national identity card or PNO - (national) personal number)
                 SemanticsIdentifier.CountryCode.EE, // 2 character ISO 3166-1 alpha-2 country code
-                "10101010005"); // identifier (according to country and identity type reference)
+                "30303039914"); // identifier (according to country and identity type reference)
 
         // For security reasons a new hash value must be created for each new authentication request
         AuthenticationHash authenticationHash = AuthenticationHash.generateRandomHash();
@@ -215,7 +209,7 @@ It is recommended to read trusted certificates from a file.
      */
 
 
-    @Test(expected = RequiredInteractionNotSupportedByAppException.class)
+    @Test
     public void documentAuthenticatingWithDocumentNumber() {
 
         AuthenticationHash authenticationHash = AuthenticationHash.generateRandomHash();
@@ -226,7 +220,7 @@ It is recommended to read trusted certificates from a file.
 
         SmartIdAuthenticationResponse authenticationResponse = client
                 .createAuthentication()
-                .withDocumentNumber("PNOEE-10101010005-Z1B2-Q")
+                .withDocumentNumber("PNOEE-30303039914-5QSV-Q")
                 .withAuthenticationHash(authenticationHash)
                 .withCertificateLevel("QUALIFIED")
                 .withAllowedInteractionsOrder(Collections.singletonList(
@@ -264,8 +258,11 @@ It is recommended to read trusted certificates from a file.
         String givenName = authIdentity.getGivenName(); // e.g. Mari-Liis"
         String surname = authIdentity.getSurname(); // e.g. "Männik"
         String identityCode = authIdentity.getIdentityNumber(); // e.g. "47101010033"
-        String country = authIdentity.getCountry(); // e.g. "EE"
+        String country = authIdentity.getCountry(); // e.g. "EE", "LV", "LT"
 
+        // Date-of-birth is extracted from certificate attribute or parsed from national identity number
+        // Value is present for all Estonian and Lithuanian persons but not for all Latvian certificates
+        Optional<LocalDate> dateOfBirth = authIdentity.getDateOfBirth();
     }
 
 
@@ -286,7 +283,7 @@ It is recommended to read trusted certificates from a file.
 
         SmartIdCertificate responseWithSigningCertificate = client
                 .getCertificate()
-                .withDocumentNumber("PNOEE-10101010005-Z1B2-Q") // returned as authentication result
+                .withDocumentNumber("PNOEE-30303039914-5QSV-Q") // returned as authentication result
                 .withCertificateLevel("QUALIFIED")
                 .fetch();
 
@@ -332,7 +329,7 @@ create the AsicE/BDoc container with files in it and get the hash to be signed.
 
         SmartIdSignature smartIdSignature = client
                 .createSignature()
-                .withDocumentNumber("PNOEE-10101010005-Z1B2-Q") // returned as authentication result
+                .withDocumentNumber("PNOEE-30303039914-5QSV-Q") // returned as authentication result
                 .withSignableHash(hashToSign)
                 .withCertificateLevel("QUALIFIED")
                 .withAllowedInteractionsOrder(asList(
@@ -381,7 +378,7 @@ Every Smart-ID app supports this interaction flow and there is no need to provid
     public void documentInteractionOrderMostCommon() {
         SmartIdSignature smartIdSignature = client
                 .createSignature()
-                .withDocumentNumber("PNOEE-10101010005-Z1B2-Q")
+                .withDocumentNumber("PNOEE-30303039914-5QSV-Q")
                 .withSignableHash(hashToSign)
                 .withCertificateLevel("QUALIFIED")
                 .withAllowedInteractionsOrder(Collections.singletonList(
@@ -408,7 +405,7 @@ If user's app doesn't support displaying verification code choice then system fa
         try {
             SmartIdSignature smartIdSignature = client
                 .createSignature()
-                .withDocumentNumber("PNOEE-10101010005-Z1B2-Q")
+                .withDocumentNumber("PNOEE-30303039914-5QSV-Q")
                 .withSignableHash(hashToSign)
                 .withCertificateLevel("QUALIFIED")
                 .withAllowedInteractionsOrder(Arrays.asList(
@@ -436,7 +433,7 @@ If the Smart-ID app in user's smart device doesn't support this feature then the
     public void documentInteractionOrderConfirmationWithFallbackToPin() {
         SmartIdSignature smartIdSignature = client
                 .createSignature()
-                .withDocumentNumber("PNOEE-10101010005-Z1B2-Q")
+                .withDocumentNumber("PNOEE-30303039914-5QSV-Q")
                 .withSignableHash(hashToSign)
                 .withCertificateLevel("QUALIFIED")
                 .withAllowedInteractionsOrder(asList(
@@ -468,7 +465,7 @@ If user picks wrong verification code then the session is cancelled and library 
     public void documentInteractionOrder2() {
         SmartIdSignature smartIdSignature = client
                 .createSignature()
-                .withDocumentNumber("PNOEE-10101010005-Z1B2-Q")
+                .withDocumentNumber("PNOEE-30303039914-5QSV-Q")
                 .withSignableHash(hashToSign)
                 .withCertificateLevel("QUALIFIED")
                 .withAllowedInteractionsOrder(asList(
@@ -504,7 +501,7 @@ If End User's phone doesn't support required flow the library throws `RequiredIn
         try {
             client
                 .createSignature()
-                .withDocumentNumber("PNOEE-10101010005-Z1B2-Q")
+                .withDocumentNumber("PNOEE-30303039914-5QSV-Q")
                 .withSignableHash(hashToSign)
                 .withCertificateLevel("QUALIFIED")
                 .withAllowedInteractionsOrder(Collections.singletonList(