+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.1] - 2021-07-07
+
+### Added
+- AuthenticationIdentity.getDateOfBirth() to get person birthdate (if available).
+- Add library version number and Java major release number to User-Agent header of outgoing requests
+
 ## [2.0] - 2020-11-20
 
 ### Changed